Split out views/data.py
