+++ corehq/apps/hqadmin/views/__init__.py
@@ -1,5 +1,13 @@
 from __future__ import absolute_import
 
+from corehq.apps.hqadmin.views.data import (
+    RecentCouchChangesView,
+    _get_db_from_db_name,
+    doc_in_es,
+    download_recent_changes,
+    raw_couch,
+    raw_doc,
+)
 from corehq.apps.hqadmin.views.system import (
     SystemInfoView,
     branches_on_staging,
@@ -15,25 +23,16 @@ from corehq.apps.hqadmin.views.views import (
     BaseAdminSectionView,
     AuthenticateAs,
     SuperuserManagement,
-    RecentCouchChangesView,
     mass_email,
-    download_recent_changes,
     AdminRestoreView,
     DomainAdminRestoreView,
     run_command,
     FlagBrokenBuilds,
     stats_data,
     admin_reports_stats_data,
-    _Db,
-    _SQL_DBS,
-    _get_db_from_db_name,
-    _lookup_id_in_database,
     web_user_lookup,
     DisableUserView,
     DisableTwoFactorView,
-    doc_in_es,
-    raw_couch,
-    raw_doc,
     callcenter_test,
     DownloadMALTView,
     _malt_csv_response,
