+++ corehq/apps/hqadmin/views/data.py
@@ -1,0 +1,294 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import unicode_literals
+
+import itertools
+import six.moves.html_parser
+import json
+import socket
+import uuid
+from io import StringIO
+from collections import defaultdict, namedtuple, OrderedDict, Counter
+from datetime import timedelta, date, datetime
+
+import dateutil
+from couchdbkit import ResourceNotFound
+from django.conf import settings
+from django.contrib import messages
+from django.contrib.auth import login
+from django.contrib.auth.models import User
+from django.core import management, cache
+from django.core.exceptions import ObjectDoesNotExist
+from django.core.mail import mail_admins
+from django.http import (
+    HttpResponseRedirect,
+    HttpResponse,
+    HttpResponseBadRequest,
+    HttpResponseNotFound,
+    JsonResponse,
+    StreamingHttpResponse,
+)
+from django.http.response import Http404
+from django.shortcuts import render, redirect
+from django.template.loader import render_to_string
+from django.utils.decorators import method_decorator
+from django.utils.functional import cached_property
+from django.utils.translation import ugettext as _, ugettext_lazy
+from django.views.decorators.http import require_POST
+from django.views.generic import FormView, TemplateView, View
+from lxml import etree
+from lxml.builder import E
+from restkit import Resource
+from restkit.errors import Unauthorized
+
+from casexml.apps.case.models import CommCareCase
+from casexml.apps.phone.xml import SYNC_XMLNS
+from casexml.apps.stock.const import COMMTRACK_REPORT_XMLNS
+from corehq.apps.app_manager.models import ApplicationBase
+from corehq.apps.callcenter.indicator_sets import CallCenterIndicators
+from corehq.apps.callcenter.utils import CallCenterCase
+from corehq.apps.data_analytics.admin import MALTRowAdmin
+from corehq.apps.data_analytics.const import GIR_FIELDS
+from corehq.apps.data_analytics.models import MALTRow, GIRRow
+from corehq.apps.domain.auth import basicauth
+from corehq.apps.domain.decorators import (
+    require_superuser, require_superuser_or_contractor,
+    login_or_basic, domain_admin_required,
+    check_lockout)
+from corehq.apps.domain.models import Domain
+from corehq.apps.es import filters
+from corehq.apps.es.domains import DomainES
+from corehq.apps.hqadmin.reporting.exceptions import HistoTypeNotFoundException
+from corehq.apps.hqadmin.service_checks import run_checks
+from corehq.apps.hqwebapp.views import BaseSectionPageView
+from corehq.apps.locations.models import SQLLocation
+from corehq.apps.ota.views import get_restore_response, get_restore_params
+from corehq.apps.hqwebapp.decorators import use_datatables, use_jquery_ui, \
+    use_nvd3_v3
+from corehq.apps.users.models import CommCareUser, WebUser, CouchUser
+from corehq.apps.users.util import format_username
+from corehq.elastic import parse_args_for_es, run_query, ES_META
+from corehq.form_processor.backends.couch.dbaccessors import CaseAccessorCouch
+from corehq.form_processor.backends.sql.dbaccessors import CaseAccessorSQL
+from corehq.form_processor.exceptions import XFormNotFound, CaseNotFound
+from corehq.form_processor.interfaces.dbaccessors import CaseAccessors
+from corehq.form_processor.models import XFormInstanceSQL, CommCareCaseSQL
+from corehq.form_processor.serializers import XFormInstanceSQLRawDocSerializer, \
+    CommCareCaseSQLRawDocSerializer
+from corehq.toggles import any_toggle_enabled, SUPPORT
+from corehq.util import reverse
+from corehq.util.couchdb_management import couch_config
+from corehq.util.supervisord.api import (
+    PillowtopSupervisorApi,
+    SupervisorException,
+    all_pillows_supervisor_status,
+    pillow_supervisor_status
+)
+from corehq.util.timer import TimingContext
+from couchforms.models import XFormInstance
+from couchforms.openrosa_response import RESPONSE_XMLNS
+from dimagi.utils.couch.database import get_db, is_bigcouch
+from dimagi.utils.csv import UnicodeWriter
+from dimagi.utils.dates import add_months
+from dimagi.utils.decorators.datespan import datespan_in_request
+from memoized import memoized
+from dimagi.utils.django.email import send_HTML_email
+from dimagi.utils.django.management import export_as_csv_action
+from dimagi.utils.parsing import json_format_date
+from dimagi.utils.web import json_response
+from corehq.apps.hqadmin.tasks import send_mass_emails
+from pillowtop.exceptions import PillowNotFoundError
+from pillowtop.utils import get_all_pillows_json, get_pillow_json, get_pillow_config_by_name
+from corehq.apps.hqadmin import service_checks, escheck
+from corehq.apps.hqadmin.forms import (
+    AuthenticateAsForm, BrokenBuildsForm, EmailForm, SuperuserManagementForm,
+    ReprocessMessagingCaseUpdatesForm,
+    DisableTwoFactorForm, DisableUserForm)
+from corehq.apps.hqadmin.history import get_recent_changes, download_changes
+from corehq.apps.hqadmin.models import HqDeploy
+from corehq.apps.hqadmin.reporting.reports import get_project_spaces, get_stats_data, HISTO_TYPE_TO_FUNC
+from corehq.apps.hqadmin.utils import get_celery_stats
+from corehq.apps.hqadmin.views.views import BaseAdminSectionView, get_hqadmin_base_context
+import six
+from six.moves import filter
+
+
+class _Db(object):
+    """
+    Light wrapper for providing interface like Couchdbkit's Database objects.
+    """
+
+    def __init__(self, dbname, getter, doc_type):
+        self.dbname = dbname
+        self._getter = getter
+        self.doc_type = doc_type
+
+    def get(self, record_id):
+        try:
+            return self._getter(record_id)
+        except (XFormNotFound, CaseNotFound, ObjectDoesNotExist):
+            raise ResourceNotFound("missing")
+
+
+_SQL_DBS = OrderedDict((db.dbname, db) for db in [
+    _Db(
+        XFormInstanceSQL._meta.db_table,
+        lambda id_: XFormInstanceSQLRawDocSerializer(XFormInstanceSQL.get_obj_by_id(id_)).data,
+        XFormInstanceSQL.__name__
+    ),
+    _Db(
+        CommCareCaseSQL._meta.db_table,
+        lambda id_: CommCareCaseSQLRawDocSerializer(CommCareCaseSQL.get_obj_by_id(id_)).data,
+        CommCareCaseSQL.__name__
+    ),
+    _Db(
+        SQLLocation._meta.db_table,
+        lambda id_: SQLLocation.objects.get(location_id=id_).to_json(),
+        SQLLocation.__name__
+    ),
+])
+
+
+def _get_db_from_db_name(db_name):
+    if db_name in _SQL_DBS:
+        return _SQL_DBS[db_name]
+    elif db_name == couch_config.get_db(None).dbname:  # primary db
+        return couch_config.get_db(None)
+    else:
+        return couch_config.get_db(db_name)
+
+
+def _lookup_id_in_database(doc_id, db_name=None):
+    db_result = namedtuple('db_result', 'dbname result status')
+    STATUSES = defaultdict(lambda: 'warning', {
+        'missing': 'default',
+        'deleted': 'danger',
+    })
+
+    response = {"doc_id": doc_id}
+    if db_name:
+        dbs = [_get_db_from_db_name(db_name)]
+        response['selected_db'] = db_name
+    else:
+        couch_dbs = list(couch_config.all_dbs_by_slug.values())
+        sql_dbs = list(_SQL_DBS.values())
+        dbs = couch_dbs + sql_dbs
+
+    db_results = []
+    for db in dbs:
+        try:
+            doc = db.get(doc_id)
+        except ResourceNotFound as e:
+            db_results.append(db_result(db.dbname, e.msg, STATUSES[e.msg]))
+        else:
+            db_results.append(db_result(db.dbname, 'found', 'success'))
+            response.update({
+                "doc": json.dumps(doc, indent=4, sort_keys=True),
+                "doc_type": doc.get('doc_type', getattr(db, 'doc_type', 'Unknown')),
+                "dbname": db.dbname,
+            })
+
+    response['db_results'] = db_results
+    return response
+
+
+@require_superuser
+def doc_in_es(request):
+    doc_id = request.GET.get("id")
+    if not doc_id:
+        return render(request, "hqadmin/doc_in_es.html", {})
+
+    def to_json(doc):
+        return json.dumps(doc, indent=4, sort_keys=True) if doc else "NOT FOUND!"
+
+    query = {"filter": {"ids": {"values": [doc_id]}}}
+    found_indices = {}
+    es_doc_type = None
+    for index in ES_META:
+        res = run_query(index, query)
+        if 'hits' in res and res['hits']['total'] == 1:
+            es_doc = res['hits']['hits'][0]['_source']
+            found_indices[index] = to_json(es_doc)
+            es_doc_type = es_doc_type or es_doc.get('doc_type')
+
+    context = {
+        "doc_id": doc_id,
+        "es_info": {
+            "status": "found" if found_indices else "NOT FOUND IN ELASTICSEARCH!",
+            "doc_type": es_doc_type,
+            "found_indices": found_indices,
+        },
+        "couch_info": _lookup_id_in_database(doc_id),
+    }
+    return render(request, "hqadmin/doc_in_es.html", context)
+
+
+@require_superuser
+def raw_couch(request):
+    get_params = dict(six.iteritems(request.GET))
+    return HttpResponseRedirect(reverse("raw_doc", params=get_params))
+
+
+@require_superuser
+def raw_doc(request):
+    doc_id = request.GET.get("id")
+    db_name = request.GET.get("db_name", None)
+    if db_name and "__" in db_name:
+        db_name = db_name.split("__")[-1]
+    context = _lookup_id_in_database(doc_id, db_name) if doc_id else {}
+
+    if request.GET.get("raw", False):
+        if 'doc' in context:
+            return HttpResponse(context['doc'], content_type="application/json")
+        else:
+            return HttpResponse(json.dumps({"status": "missing"}),
+                                content_type="application/json", status=404)
+
+    other_couch_dbs = sorted([_f for _f in couch_config.all_dbs_by_slug if _f])
+    context['all_databases'] = ['commcarehq'] + other_couch_dbs + list(_SQL_DBS)
+    context['use_code_mirror'] = request.GET.get('code_mirror', 'true').lower() == 'true'
+    return render(request, "hqadmin/raw_couch.html", context)
+
+
+class RecentCouchChangesView(BaseAdminSectionView):
+    urlname = 'view_recent_changes'
+    template_name = 'hqadmin/couch_changes.html'
+    page_title = ugettext_lazy("Recent Couch Changes")
+
+    @use_nvd3_v3
+    @use_datatables
+    @use_jquery_ui
+    @method_decorator(require_superuser_or_contractor)
+    def dispatch(self, *args, **kwargs):
+        return super(RecentCouchChangesView, self).dispatch(*args, **kwargs)
+
+    @property
+    def page_context(self):
+        count = int(self.request.GET.get('changes', 1000))
+        changes = list(get_recent_changes(get_db(), count))
+        domain_counts = defaultdict(lambda: 0)
+        doc_type_counts = defaultdict(lambda: 0)
+        for change in changes:
+            domain_counts[change['domain']] += 1
+            doc_type_counts[change['doc_type']] += 1
+
+        def _to_chart_data(data_dict):
+            return [
+                {'label': l, 'value': v} for l, v in sorted(list(data_dict.items()), key=lambda tup: tup[1], reverse=True)
+            ][:20]
+
+        return {
+            'count': count,
+            'recent_changes': changes,
+            'domain_data': {'key': 'domains', 'values': _to_chart_data(domain_counts)},
+            'doc_type_data': {'key': 'doc types', 'values': _to_chart_data(doc_type_counts)},
+        }
+
+
+@require_superuser_or_contractor
+def download_recent_changes(request):
+    count = int(request.GET.get('changes', 10000))
+    resp = HttpResponse(content_type='text/csv')
+    resp['Content-Disposition'] = 'attachment; filename="recent_changes.csv"'
+    download_changes(get_db(), count, resp)
+    return resp
