+++ corehq/apps/hqadmin/views/views.py
@@ -216,41 +216,6 @@ class SuperuserManagement(BaseAdminSectionView):
         return self.get(request, *args, **kwargs)
 
 
-class RecentCouchChangesView(BaseAdminSectionView):
-    urlname = 'view_recent_changes'
-    template_name = 'hqadmin/couch_changes.html'
-    page_title = ugettext_lazy("Recent Couch Changes")
-
-    @use_nvd3_v3
-    @use_datatables
-    @use_jquery_ui
-    @method_decorator(require_superuser_or_contractor)
-    def dispatch(self, *args, **kwargs):
-        return super(RecentCouchChangesView, self).dispatch(*args, **kwargs)
-
-    @property
-    def page_context(self):
-        count = int(self.request.GET.get('changes', 1000))
-        changes = list(get_recent_changes(get_db(), count))
-        domain_counts = defaultdict(lambda: 0)
-        doc_type_counts = defaultdict(lambda: 0)
-        for change in changes:
-            domain_counts[change['domain']] += 1
-            doc_type_counts[change['doc_type']] += 1
-
-        def _to_chart_data(data_dict):
-            return [
-                {'label': l, 'value': v} for l, v in sorted(list(data_dict.items()), key=lambda tup: tup[1], reverse=True)
-            ][:20]
-
-        return {
-            'count': count,
-            'recent_changes': changes,
-            'domain_data': {'key': 'domains', 'values': _to_chart_data(domain_counts)},
-            'doc_type_data': {'key': 'doc types', 'values': _to_chart_data(doc_type_counts)},
-        }
-
-
 @require_superuser_or_contractor
 def mass_email(request):
     if request.method == "POST":
@@ -271,15 +236,6 @@ def mass_email(request):
     context['hide_filters'] = True
     context['form'] = form
     return render(request, "hqadmin/mass_email.html", context)
-
-
-@require_superuser_or_contractor
-def download_recent_changes(request):
-    count = int(request.GET.get('changes', 10000))
-    resp = HttpResponse(content_type='text/csv')
-    resp['Content-Disposition'] = 'attachment; filename="recent_changes.csv"'
-    download_changes(get_db(), count, resp)
-    return resp
 
 
 class AdminRestoreView(TemplateView):
@@ -510,85 +466,6 @@ def admin_reports_stats_data(request):
     return stats_data(request)
 
 
-class _Db(object):
-    """
-    Light wrapper for providing interface like Couchdbkit's Database objects.
-    """
-
-    def __init__(self, dbname, getter, doc_type):
-        self.dbname = dbname
-        self._getter = getter
-        self.doc_type = doc_type
-
-    def get(self, record_id):
-        try:
-            return self._getter(record_id)
-        except (XFormNotFound, CaseNotFound, ObjectDoesNotExist):
-            raise ResourceNotFound("missing")
-
-
-_SQL_DBS = OrderedDict((db.dbname, db) for db in [
-    _Db(
-        XFormInstanceSQL._meta.db_table,
-        lambda id_: XFormInstanceSQLRawDocSerializer(XFormInstanceSQL.get_obj_by_id(id_)).data,
-        XFormInstanceSQL.__name__
-    ),
-    _Db(
-        CommCareCaseSQL._meta.db_table,
-        lambda id_: CommCareCaseSQLRawDocSerializer(CommCareCaseSQL.get_obj_by_id(id_)).data,
-        CommCareCaseSQL.__name__
-    ),
-    _Db(
-        SQLLocation._meta.db_table,
-        lambda id_: SQLLocation.objects.get(location_id=id_).to_json(),
-        SQLLocation.__name__
-    ),
-])
-
-
-def _get_db_from_db_name(db_name):
-    if db_name in _SQL_DBS:
-        return _SQL_DBS[db_name]
-    elif db_name == couch_config.get_db(None).dbname:  # primary db
-        return couch_config.get_db(None)
-    else:
-        return couch_config.get_db(db_name)
-
-
-def _lookup_id_in_database(doc_id, db_name=None):
-    db_result = namedtuple('db_result', 'dbname result status')
-    STATUSES = defaultdict(lambda: 'warning', {
-        'missing': 'default',
-        'deleted': 'danger',
-    })
-
-    response = {"doc_id": doc_id}
-    if db_name:
-        dbs = [_get_db_from_db_name(db_name)]
-        response['selected_db'] = db_name
-    else:
-        couch_dbs = list(couch_config.all_dbs_by_slug.values())
-        sql_dbs = list(_SQL_DBS.values())
-        dbs = couch_dbs + sql_dbs
-
-    db_results = []
-    for db in dbs:
-        try:
-            doc = db.get(doc_id)
-        except ResourceNotFound as e:
-            db_results.append(db_result(db.dbname, e.msg, STATUSES[e.msg]))
-        else:
-            db_results.append(db_result(db.dbname, 'found', 'success'))
-            response.update({
-                "doc": json.dumps(doc, indent=4, sort_keys=True),
-                "doc_type": doc.get('doc_type', getattr(db, 'doc_type', 'Unknown')),
-                "dbname": db.dbname,
-            })
-
-    response['db_results'] = db_results
-    return response
-
-
 @require_superuser
 def web_user_lookup(request):
     template = "hqadmin/web_user_lookup.html"
@@ -783,64 +660,6 @@ class DisableTwoFactorView(FormView):
 
         messages.success(self.request, _('Two-Factor Auth successfully disabled.'))
         return redirect('{}?q={}'.format(reverse('web_user_lookup'), username))
-
-
-@require_superuser
-def doc_in_es(request):
-    doc_id = request.GET.get("id")
-    if not doc_id:
-        return render(request, "hqadmin/doc_in_es.html", {})
-
-    def to_json(doc):
-        return json.dumps(doc, indent=4, sort_keys=True) if doc else "NOT FOUND!"
-
-    query = {"filter": {"ids": {"values": [doc_id]}}}
-    found_indices = {}
-    es_doc_type = None
-    for index in ES_META:
-        res = run_query(index, query)
-        if 'hits' in res and res['hits']['total'] == 1:
-            es_doc = res['hits']['hits'][0]['_source']
-            found_indices[index] = to_json(es_doc)
-            es_doc_type = es_doc_type or es_doc.get('doc_type')
-
-    context = {
-        "doc_id": doc_id,
-        "es_info": {
-            "status": "found" if found_indices else "NOT FOUND IN ELASTICSEARCH!",
-            "doc_type": es_doc_type,
-            "found_indices": found_indices,
-        },
-        "couch_info": _lookup_id_in_database(doc_id),
-    }
-    return render(request, "hqadmin/doc_in_es.html", context)
-
-
-@require_superuser
-def raw_couch(request):
-    get_params = dict(six.iteritems(request.GET))
-    return HttpResponseRedirect(reverse("raw_doc", params=get_params))
-
-
-@require_superuser
-def raw_doc(request):
-    doc_id = request.GET.get("id")
-    db_name = request.GET.get("db_name", None)
-    if db_name and "__" in db_name:
-        db_name = db_name.split("__")[-1]
-    context = _lookup_id_in_database(doc_id, db_name) if doc_id else {}
-
-    if request.GET.get("raw", False):
-        if 'doc' in context:
-            return HttpResponse(context['doc'], content_type="application/json")
-        else:
-            return HttpResponse(json.dumps({"status": "missing"}),
-                                content_type="application/json", status=404)
-
-    other_couch_dbs = sorted([_f for _f in couch_config.all_dbs_by_slug if _f])
-    context['all_databases'] = ['commcarehq'] + other_couch_dbs + list(_SQL_DBS)
-    context['use_code_mirror'] = request.GET.get('code_mirror', 'true').lower() == 'true'
-    return render(request, "hqadmin/raw_couch.html", context)
 
 
 @require_superuser
